requirements.txt
keras
tensorflow
numpy
scikit-learn
Pillow
jsmin
